refactor(example): add barrel import to models
