+++ projects/example-app/src/app/auth/containers/login-page.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Store, select } from '@ngrx/store';
-import { Credentials } from '@example-app/auth/models/user';
+import { Credentials } from '@example-app/auth/models';
 import * as fromAuth from '@example-app/auth/reducers';
 import { LoginPageActions } from '@example-app/auth/actions';
 
@@ -10,7 +10,8 @@ import { LoginPageActions } from '@example-app/auth/actions';
     <bc-login-form
       (submitted)="onSubmit($event)"
       [pending]="pending$ | async"
-      [errorMessage]="error$ | async">
+      [errorMessage]="error$ | async"
+    >
     </bc-login-form>
   `,
   styles: [],
