+++ projects/example-app/src/app/auth/effects/auth.effects.spec.ts
@@ -11,7 +11,7 @@ import {
   AuthApiActions,
 } from '@example-app/auth/actions';
 
-import { Credentials, User } from '@example-app/auth/models/user';
+import { Credentials, User } from '@example-app/auth/models';
 import { AuthService } from '@example-app/auth/services';
 import { AuthEffects } from '@example-app/auth/effects';
