+++ projects/example-app/src/app/auth/effects/auth.effects.ts
@@ -9,7 +9,7 @@ import {
   AuthActions,
   AuthApiActions,
 } from '@example-app/auth/actions';
-import { Credentials } from '@example-app/auth/models/user';
+import { Credentials } from '@example-app/auth/models';
 import { AuthService } from '@example-app/auth/services';
 import { LogoutConfirmationDialogComponent } from '@example-app/auth/components';
 
@@ -60,11 +60,8 @@ export class AuthEffects {
 
         return dialogRef.afterClosed();
       }),
-      map(
-        result =>
-          result
-            ? AuthActions.logout()
-            : AuthActions.logoutConfirmationDismiss()
+      map(result =>
+        result ? AuthActions.logout() : AuthActions.logoutConfirmationDismiss()
       )
     )
   );
