+++ projects/example-app/src/app/auth/models/index.ts
@@ -1,0 +1,1 @@
+export * from './user';
