+++ projects/example-app/src/app/auth/services/auth.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable, of, throwError } from 'rxjs';
 
-import { Credentials, User } from '@example-app/auth/models/user';
+import { Credentials, User } from '@example-app/auth/models';
 
 @Injectable({
   providedIn: 'root',
